(#80) support dependencies in tasks
This introduce the notion of dependencies between jobs
a job with dependencies starts off as Blocked and will
periodically check on it's dependencies, if its dependencies
fails it becomes Unreachable

This adjusts the default queue retries up to unlimited
instead defaulting task tries to 10, end result more or
less the same but the change it needed because checking
a blocked task dependencies is a delivery attempt on
the queue, so it would prematurely terminate blocked tasks

This is in aid of supporting a tool that can create a large
set of tasks, perhaps from a DAG, and have the processors
handle the exection.

At present it can take some time to process a large DAG of
tasks since each one will individiaully - every 5 seconds -
decide if its Unreachable or not and it only checks its
immediate parents, so an early failure on a long chain of
tasks can take some time to bubble through, will think about
optimising that. The lack of transactions for KV is a pain.

Signed-off-by: R.I.Pienaar <rip@devco.net>